msm8974_pwm_vibrator: Fix the negative remaining timeout
It's possible that hrtimer_get_remaining() returns the nagative value.
It causes the unintended behavior in the delayed workqueue scheduling.
So if the value is native, it means the timer has been already expired.
We just set the delay to 0 in this case.

And use the monotonic timer for the time stamp.

Change-Id: Ia16433d6c2715875f2d6f9da14e8e39824c15aa7
Signed-off-by: Devin Kim <dojip.kim@lge.com>
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>